+++ src/main/kotlin/io/github/durun/nitron/core/ast/node/AstRuleNode.kt
@@ -1,5 +1,6 @@
 package io.github.durun.nitron.core.ast.node
 
+import io.github.durun.nitron.core.ast.type.RuleType
 import io.github.durun.nitron.core.ast.visitor.AstVisitor
 
 /**
