+++ src/main/kotlin/io/github/durun/nitron/core/ast/type/NodeType.kt
@@ -1,0 +1,9 @@
+package io.github.durun.nitron.core.ast.node
+
+
+interface NodeType : Map.Entry<Int, String> {
+    val index: Int
+        get() = this.key
+    val name: String
+        get() = this.value
+}
