+++ src/main/kotlin/io/github/durun/nitron/core/ast/type/RuleType.kt
@@ -1,0 +1,19 @@
+package io.github.durun.nitron.core.ast.type
+
+import io.github.durun.nitron.core.ast.node.NodeType
+import io.github.durun.nitron.util.SimpleEntry
+
+class RuleType private constructor(
+        private val entry: Map.Entry<Int, String>
+) : NodeType,
+        Map.Entry<Int, String> by entry {
+
+    constructor(index: Int, name: String) : this(SimpleEntry(index, name))
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        return (other is RuleType) && (key == other.key)
+    }
+
+    override fun hashCode(): Int = key
+}
