+++ src/main/kotlin/io/github/durun/nitron/core/ast/visitor/AstSplitVisitor.kt
@@ -2,6 +2,7 @@ package io.github.durun.nitron.core.ast.visitor
 
 import io.github.durun.nitron.core.InvalidTypeException
 import io.github.durun.nitron.core.ast.node.*
+import io.github.durun.nitron.core.ast.type.NodeTypePool
 
 fun astSplitVisitorOf(splitTypes: List<String>): AstSplitVisitor {
     return StringAstSplitVisitor(splitTypes)
