+++ src/main/kotlin/io/github/durun/nitron/core/parser/AstBuildVisitor.kt
@@ -4,7 +4,7 @@ import io.github.durun.nitron.core.antlr4util.children
 import io.github.durun.nitron.core.ast.node.AstNode
 import io.github.durun.nitron.core.ast.node.AstTerminalNode
 import io.github.durun.nitron.core.ast.node.BasicAstRuleNode
-import io.github.durun.nitron.core.ast.node.nodeTypePoolOf
+import io.github.durun.nitron.core.ast.type.nodeTypePoolOf
 import org.antlr.v4.runtime.Parser
 import org.antlr.v4.runtime.tree.*
